+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+- (Changed) ``Statechart.copy_from_statechart`` has only its first argument that can be provided by position.
+  The remaining ones (esp. ``source`` and ``replace``) should be provided by name.
+
+
 0.26.3 (2018-03-15)
 -------------------